src/pyproject.toml
[project]
name = "vercel-fastapi-starter"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.121.0",
  "uvicorn>=0.34.0",
  "SQLAlchemy",
  "psycopg2-binary",
  "pydantic",
  "pydantic[email]",
  "httpx",
  "alembic",
  "bcrypt",
  "PyJWT",
  "python-decouple",
]

[project.optional-dependencies]
dev = [
  "pytest",
]
desktop = [
  "pyside6",
]
